+++ backend/src/main/java/com/jujeol/commons/interceptor/LogInterceptor.java
@@ -1,0 +1,26 @@
+package com.jujeol.commons.interceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.ModelAndView;
+
+public class LogInterceptor implements HandlerInterceptor {
+
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response,
+            Object handler) {
+        logger.info("----------------------- \"{}\" Start -----------------------", request.getRequestURI());
+        return true;
+    }
+
+    @Override
+    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler,
+            ModelAndView modelAndView) {
+        logger.info("----------------------- \"{}\" End -----------------------", request.getRequestURI());
+    }
+}
